Scheduled recordings (#59)
* Add full API support for Recording Schedules

* Add Recording Scheduling backend logic

* UI: Add basic logic for handling with schedules

* Progress - Schedules table

* Finish schedules table, remove scheduled-brief object

* Add Schedule options in device recorder screen

* Fix: Load Recording schedules on startup

* version

+++ web_service/src/Scheduler/ConnectionScheduler.cs
@@ -39,7 +39,7 @@ public class ConnectionScheduler : TaskScheduler{
 
     #endregion
 
-    private const int SCHEDULING_PERIOD_MINUTES = 2;
+    private const double SCHEDULING_PERIOD_MINUTES = 2;
 
     private void ConnectRadars()
     {
